move console output methods from AbstractView to new ConsoleOutputHandler class
