+++ src/main/java/com/worldwarofants/game/arch/AbstractView.java
@@ -1,7 +1,5 @@
 package com.worldwarofants.game.arch;
 
-import java.util.concurrent.TimeUnit;
-
 /**
  * Views are the classes responsible for interacting with the console.
  * They mainly print and read.
@@ -15,7 +13,6 @@ import java.util.concurrent.TimeUnit;
 public abstract class AbstractView<VM extends IViewModel> {
 
     private VM viewModel;
-    private static long spellDelay = 100;
 
     public AbstractView(VM viewModel) {
         this.viewModel = viewModel;
@@ -23,28 +20,5 @@ public abstract class AbstractView<VM extends IViewModel> {
 
     public VM getViewModel() {
         return viewModel;
-    }
-
-    //output the provided string in one line, character by character
-    public static void spell(String input){
-        for(int i = 0; i < input.length(); i++){
-            System.out.print(input.charAt(i));
-            try{
-                TimeUnit.MILLISECONDS.sleep(spellDelay);
-            }
-            catch(Exception e){
-                e.printStackTrace();
-            }
-        }
-    }
-    
-    //output the provided string in one line
-    public static void post(String input){
-        System.out.println(input);   
-    }
-
-    //output a line break
-    public static void line_break(){
-        System.out.println(); 
     }
 }
